requirements.txt
pyyaml
numpy
opencv-python-headless
python-pptx
webcolors
fastapi
uvicorn[standard]
reportlab